# docs/api-documentation.yaml
openapi: 3.0.3
info:
  title: Mobile Travel App API
  description: |
    Comprehensive REST API for Mobile Travel App Backend System

    ## Features
    - **User Management**: Admins, Customers, and Drivers
    - **Coin System**: Top-up requests and transaction tracking
    - **Route & Vehicle Management**: Travel routes and vehicle fleet
    - **Schedule Management**: Trip scheduling with conflict detection
    - **Ticket Booking**: Passenger booking with seat selection
    - **Travel Documents**: Digital travel permits with PDF generation
    - **Driver Panel**: Real-time trip tracking and passenger management
    - **Trip Logs**: Complete audit trail of all trips

    ## Authentication
    All endpoints require JWT Bearer token authentication unless marked as public.

    ## Coin System
    - Ticket booking: 10,000 coins per passenger
    - Travel document: 10,000 coins per document

  version: 1.0.0
  contact:
    name: API Support
    email: support@travelapp.com
    url: https://github.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Local Development
  - url: https://api.travelapp.com/api/v1
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
  schemas:
...
    UserRole:
      type: string
      enum: [SUPER_ADMIN, ADMIN, DRIVER, CUSTOMER]
    UserStatus:
      type: string
      enum: [ACTIVE, INACTIVE, SUSPENDED]

    # Coin System
...
    ScheduleStatus:
      type: string
      enum: [SCHEDULED, DEPARTED, ARRIVED, CANCELLED]
...
    TravelDocumentStatus:
      type: string
      enum: [DRAFT, ISSUED, CANCELLED]

    # Trip Status
    TripStatus:
      type: string
      enum: [ASSIGNED, READY, DEPARTED, IN_TRANSIT, REST_STOP, ARRIVED, COMPLETED, CANCELLED]
    DriverStatus:
      type: string
      enum: [AVAILABLE, ON_TRIP, OFF_DUTY]

    # Error Response
    Error:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: Validation failed
        error:
          type: string
          example: Bad Request

    # Pagination
...
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user with phone number and receive JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, password]
              properties:
                phone:
                  type: string
                  pattern: '^(\+62|62|0)[0-9]{9,12}$'
                  example: "081234567890"
                  description: Indonesian phone number format
                password:
                  type: string
                  format: password
                  minLength: 6
                  example: Admin123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      phone:
                        type: string
                      email:
                        type: string
                        nullable: true
                      role:
                        $ref: '#/components/schemas/UserRole'
                      status:
                        $ref: '#/components/schemas/UserStatus'
                      profile:
                        type: object
                        nullable: true
                        properties:
                          id:
                            type: string
                          name:
                            type: string
                          phone:
                            type: string
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/register:
    post:
      tags: [Authentication]
      summary: Customer registration
      description: Register new customer account with phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, phone, password]
              properties:
                name:
                  type: string
                  minLength: 3
                  example: John Doe
                phone:
                  type: string
                  pattern: '^(\+62|62|0)[0-9]{9,12}$'
                  example: "081234567890"
                  description: Indonesian phone number format
                email:
                  type: string
                  format: email
                  example: john@example.com
                  description: Optional email address
                password:
                  type: string
                  format: password
                  minLength: 6
                  pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)'
                  example: Customer123
                  description: Must contain uppercase, lowercase, and number
                address:
                  type: string
                  example: Jl. Sudirman No. 123, Jakarta
                  description: Optional address
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      phone:
                        type: string
                      email:
                        type: string
                        nullable: true
                      role:
                        $ref: '#/components/schemas/UserRole'
                      status:
                        $ref: '#/components/schemas/UserStatus'
                      profile:
                        type: object
                        properties:
                          id:
                            type: string
                          name:
                            type: string
                          phone:
                            type: string
        '409':
          description: Phone number or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: Get new access token using refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      phone:
                        type: string
                      email:
                        type: string
                        nullable: true
                      role:
                        $ref: '#/components/schemas/UserRole'
                      status:
                        $ref: '#/components/schemas/UserStatus'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get current user profile
      description: Retrieve authenticated user's profile information
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  phone:
                    type: string
                  email:
                    type: string
                    nullable: true
                  role:
                    $ref: '#/components/schemas/UserRole'
                  status:
                    $ref: '#/components/schemas/UserStatus'
                  profile:
                    type: object
                    nullable: true
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      phone:
                        type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== COIN SYSTEM ====================
...
  /travel-documents:
    post:
      tags: [Travel Documents]
      summary: Create travel document
      description: Create draft travel document (Surat Jalan)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scheduleId, vehicleId, driverName, driverPhone, totalPassengers, departureDate]
              properties:
                scheduleId:
                  type: string
                vehicleId:
                  type: string
                driverName:
                  type: string
                  example: Ahmad Supardi
                driverPhone:
                  type: string
                  example: '081234567890'
                totalPassengers:
                  type: integer
                  minimum: 1
                  example: 12
                departureDate:
                  type: string
                  format: date-time
                notes:
                  type: string
      responses:
        '201':
          description: Draft document created
        '400':
          description: Validation failed

    get:
      tags: [Travel Documents]
      summary: List travel documents
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TravelDocumentStatus'
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of documents
  /travel-documents/{id}/issue:
    patch:
      tags: [Travel Documents]
      summary: Issue travel document
      description: |
        Issue draft document with automatic coin deduction (10,000/document)
        - Validates coin balance
        - Deducts 10,000 coins
        - Marks as ISSUED (final state)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Document issued and coins deducted
        '400':
          description: Insufficient coin balance
  /travel-documents/{id}/print:
    get:
      tags: [Travel Documents]
      summary: Generate PDF
      description: Generate and download travel document as PDF
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary

  # ==================== DRIVER PANEL ====================
  /driver/profile:
    get:
      tags: [Driver Panel]
      summary: Get driver profile
      description: Get authenticated driver's profile and statistics
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Driver profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  phone:
                    type: string
                  licenseNumber:
                    type: string
                  status:
                    $ref: '#/components/schemas/DriverStatus'
                  _count:
                    type: object
                    properties:
                      schedules:
                        type: integer
                      tripLogs:
                        type: integer
  /driver/trips:
    get:
      tags: [Driver Panel]
      summary: Get assigned trips
      description: Get all trips assigned to authenticated driver
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ScheduleStatus'
      responses:
        '200':
          description: List of assigned trips
  /driver/trips/{scheduleId}:
    get:
      tags: [Driver Panel]
      summary: Get trip details
      description: Get complete trip information
      security:
        - BearerAuth: []
      parameters:
        - name: scheduleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Trip details
        '400':
          description: Not assigned to this trip
  /driver/trips/{scheduleId}/passengers:
    get:
      tags: [Driver Panel]
      summary: Get passenger list
      description: Get boarding manifest for trip
      security:
        - BearerAuth: []
      parameters:
        - name: scheduleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Passenger manifest
          content:
            application/json:
              schema:
                type: object
                properties:
                  schedule:
                    type: object
                  passengers:
                    type: array
                    items:
                      type: object
                      properties:
                        passengerId:
                          type: string
                        passengerName:
                          type: string
                        identityNumber:
                          type: string
                        phone:
                          type: string
                        seatNumber:
                          type: string
                        ticketNumber:
                          type: string
                        customerName:
                          type: string
                  summary:
                    type: object
                    properties:
                      totalTickets:
                        type: integer
                      totalPassengers:
                        type: integer
  /driver/trips/{scheduleId}/status:
    post:
      tags: [Driver Panel]
      summary: Update trip status
      description: |
        Update trip status with location tracking
        - Creates trip log entry
        - Auto-updates schedule status
        - Supports GPS coordinates
      security:
        - BearerAuth: []
      parameters:
        - name: scheduleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: '#/components/schemas/TripStatus'
                location:
                  type: string
                  example: Rest Area KM 45
                latitude:
                  type: number
                  example: -6.200000
                longitude:
                  type: number
                  example: 106.816666
                notes:
                  type: string
                  example: Taking 15 minute break
      responses:
        '200':
          description: Trip status updated and logged
security:
  - BearerAuth: []